waveform types changed from short to int, so both the number of samples and the amplitudes can go up to 65535

+++ emodel/Waveform.cc
@@ -18,17 +18,17 @@ gate::Waveform::~Waveform() {
 }
 
 //=======================================================
-void gate::Waveform::SetData(std::vector< std::pair<unsigned short,unsigned short> >data){
+void gate::Waveform::SetData(std::vector< std::pair<unsigned int,unsigned int> >data){
 //=======================================================
     _data=data;}
 
 //=======================================================
-double gate::Waveform::GetAmplitude(unsigned short isamp) const{
+double gate::Waveform::GetAmplitude(unsigned int isamp) const{
 //=======================================================
   
   double amp = gate::nan;
 
-  std::vector< std::pair<unsigned short,unsigned short> >::const_iterator I;
+  std::vector< std::pair<unsigned int,unsigned int> >::const_iterator I;
   
   for (I = _data.begin(); I !=_data.end(); ++I ){
